Completed valid_in_box(), is_valid(), fill_box(), fill_diagonal(). Bug in fill_remaining() fixed.
There was a bug in fill_remaining() which was in the starter code. It passed floats into col due to box_length being a float, but it is supposed to be an integer, and it was causing errors. The starter code says "DO NOT CHANGE" but I felt in this case it's probably fine.